chore: rename SubjectTOC component
