+++ solution/ui/regulations/eregs-vite/src/components/subjects/SubjectLanding.vue
